requirements.txt
flask
pillow
requests
openai
reportlab
fal-client
python-dotenv